global_private_messenger/requirements.txt
# Web Messenger Requirements

# Flask and Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# WebSocket
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0

# Security
Werkzeug==3.0.1
PyJWT==2.8.1
bcrypt==4.1.1

# Server
gunicorn==21.2.0
eventlet==0.33.3

# Utilities
requests==2.31.0
six==1.16.0

# Development (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
